aws/lambda/requirements.txt
boto3==1.35.91
botocore==1.35.91
jmespath==1.0.1
pillow==11.1.0
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.17.0
urllib3==2.3.0
